requirements.txt
Flask
Flask-Security
bcrypt
psycopg2-binary
PyMySQL
flask_marshmallow
flask_sqlalchemy
marshmallow-sqlalchemy
flask_cors
Flask-Migrate
flask_session
Flask_JWT_Extended
Flask_Bcrypt
boto3
Flask-Mail
flask-dropzone
